fix: Move to Upload function
